requirements.txt
streamlit
pandas
openpyxl
google-ortools